requirements.txt
pywebview
pandas
numpy
scikit-learn
google-generativeai
openpyxl
PyYAML
matplotlib
seaborn